RequestFulfillers now use "Supplied" ResourceBlock
adds a new field to MarketActiveRequests that lists the "supplied" resource block - i.e. what was consumed from the supplier. This field is now used to populate RequestFulfillers instead of the "requested" ResourceBlock. This should only really affect digested fertilizer, where it should now be shipped properly instead of reverting to manure